Update Date serialization to be more consistent
Before the default Date serialization was to as a number or a formatted string, with special support for Object property to serialize as { class: __Timestamp__, value: #### }.

Now we always serialize as { class: __Timestamp__ }... however Date properties contains a custom toJSON to only serialize as the number.

This means that serialization should work out of box no matter where we do it, and Date properties should continue to be efficient.

This also adds an End to End JSON Serialization Test.  The test makes an object iwth most of the known serializable properties, goes to/from JSON once on the client, send the object to the server, goes to/from JSON on the server, send the object back to the client, and then compares that result with the initial object.

Currently this test fails because of disagreement in the Server code on whether a Class property should hold the real Class object or a ClassInfo object.

But the Date portion succeeds.